requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
reportlab==4.0.8
python-docx==1.1.0
openai==1.12.0
python-dotenv==1.0.1
